+++ CHANGELOG.md
@@ -4,18 +4,41 @@
 
 # Changelog
 
-## 9.0.0 - UNRELEASED
+## 9.2.0
+
+##### Features
+- new placeholder
+  - `${build.timestamp}` maven-build-timestamp (epoch seconds) e.g. '1560694278'
+  - `${build.timestamp.year}` maven-build-timestamp year e.g. '2021'
+  - `${build.timestamp.year.2digit}` 2-digit maven-build-timestamp year.g. '21'
+  - `${build.timestamp.month}` maven-build-timestamp month of year e.g. '12'
+  - `${build.timestamp.day}` maven-build-timestamp day of month e.g. '23'
+  - `${build.timestamp.hour}` maven-build-timestamp hour of day (24h)e.g. '13'
+  - `${build.timestamp.minute}` maven-build-timestamp minute of hour e.g. '59'
+  - `${build.timestamp.second}` maven-build-timestamp second of minute e.g. '30'
+  - `${build.timestamp.datetime}` maven-build-timestamp formatted as `yyyyMMdd.HHmmss`e.g. '20190616.161442'
+
+  
+## 9.1.0
+
+##### Features
+- add config option for `<projectVersionPattern>` to use special parts of the project version as placeholders e.g. `${version.environment}`
+- add placeholder `${version.core}` the core version component of `${version}` e.g. '1.2.3'
+  - `${version.release}` is marked as deprecated
+
+
+## 9.0.1
+
+##### Fixes
+- handle lightweight tags
+- set `${version.minor.next}` placeholders properly
+
+
+## 9.0.0 (accidentally increased major version, D'OH )
 ##### Fixes
 * fix git describe tag selection, if multiple tags point to head
+* add missing `${version.label.prefixed}` placeholder
 
-##### BREAKING CHANGES
-* remove following project properties. Make use of properties configuration in config file instead.
-  * `git.commit` 
-  * `git.commit.short` 
-  * `git.commit.timestamp` 
-  * `git.commit.timestamp.datetime` 
-  * `git.ref` 
-  * `git.ref.slug` 
 
 ## 8.0.1
 ##### Fixes